split provider from hook for user state
